raft: made the fast path in LeaderTransfer easier to understand
Previously, the rp parameter provided to handleLeaderTransfer() is the
rp of the target (not the source node), as we have the From field of the
incoming message set to the transfer target in peer.go. This can make it
pretty confusing when just reading the raft.go file as all other similar
methods have the rp parameter provided as the rp of the source node.