transport: removed all mentions of the term RPC
Raft paper uses the term RPC for the transport module used in a Raft
system. That is how dragonboat used the term RPC from the beginning. The
problem is the transport module is pretty different from a tranditional
RPC setup. This caused quite a lot of confusions for users. Renamed all
"RPC" to "Transport".