Remove the CircuitBreaker for address.resolve
Using the same CircuitBreaker leads to problems if
gossip is used. If gossip is used it might be that one
node is not resolving because it is down. If this happens
the only CircuitBreaker will break sends to other available
nodes of the cluster. As as result the cluster becomes
unavailable because communication to all nodes randomly
fail and quorum or leader election can't be fulfilled anymore.