raft: introduce DisableProposalForwarding option
DisableProposalForwarding set to true means that followers will drop
proposals, rather than forwarding them to the leader. Proposal from
follower or observer will be dropped.

One use case for this feature would be in a situation where the Raft leader
is used to compute the data of a proposal, for example, adding a timestamp
from a hybrid logical clock to data in a monotonically increasing way.
Forwarding should be disabled to prevent a follower with an inaccurate hybrid
logical clock from assigning the timestamp and then forwarding the data
to the leader.